import and source code cleanup
git-svn-id: https://findbugs.googlecode.com/svn/trunk@12736 eae3c2d3-9b19-0410-a86e-396b6ccb6ab3

+++ findbugs/src/java/edu/umd/cs/findbugs/detect/FieldItemSummary.java
@@ -134,7 +134,7 @@ public class FieldItemSummary extends OpcodeStackDetector implements NonReportin
                     else if (firstChar == 'I')
                         item = new OpcodeStack.Item("I", (Integer) 0);
                     else if (firstChar == 'J')
-                        item = new OpcodeStack.Item("J", (Long) 0L);
+                        item = new OpcodeStack.Item("J", 0L);
                     else
                         item = new OpcodeStack.Item(f.getSignature());
                     fieldSummary.mergeSummary(f, item);
